reverted nested functionality, broke unit tests
